standardize naming conventions and merge files
