+++ internal/cqueue/cqueue.go
@@ -19,11 +19,65 @@
 package cqueue
 
 import (
+	"CraneFrontEnd/generated/protos"
 	"CraneFrontEnd/internal/util"
+	"context"
 	"fmt"
 	"strconv"
 	"time"
 )
+
+var (
+	stub protos.CraneCtldClient
+)
+
+func FillReqByCobraFlags() (*protos.QueryTasksInfoRequest, error) {
+	req := protos.QueryTasksInfoRequest{OptionIncludeCompletedTasks: false}
+
+	processors := []FilterProcessor{
+		&StatesFilterProcessor{},
+		&SelfProcessor{},
+		&JobNamesProcessor{},
+		&UserFilterProcessor{},
+		&QosProcessor{},
+		&AccountProcessor{},
+		&PartitionsProcessor{},
+		&JobIDsProcessor{},
+	}
+
+	for _, p := range processors {
+		if err := p.Process(&req); err != nil {
+			return nil, err
+		}
+	}
+
+	if FlagNumLimit != 0 {
+		req.NumLimit = FlagNumLimit
+	}
+
+	return &req, nil
+}
+
+func QueryTasksInfo() (*protos.QueryTasksInfoReply, error) {
+	config := util.ParseConfig(FlagConfigFilePath)
+	stub = util.GetStubToCtldByConfig(config)
+
+	req, err := FillReqByCobraFlags()
+	if err != nil {
+		return &protos.QueryTasksInfoReply{}, err
+	}
+	// Use a 10-second timeout for the RPC
+	ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
+	defer cancel()
+
+	reply, err := stub.QueryTasksInfo(ctx, req)
+	if err != nil {
+		util.GrpcErrorPrintf(err, "Failed to query task queue")
+		return nil, &util.CraneError{Code: util.ErrorNetwork}
+	}
+
+	return reply, nil
+}
 
 func Query() error {
 	reply, err := QueryTasksInfo()
@@ -39,7 +93,7 @@ func Query() error {
 func loopedQuery(iterate uint64) error {
 	interval, err := time.ParseDuration(strconv.FormatUint(iterate, 10) + "s")
 	if err != nil {
-		return util.GetCraneError(util.ErrorCmdArg, "Invalid time interval.")
+		return util.NewCraneErr(util.ErrorCmdArg, "Invalid time interval.")
 	}
 	return loopedSubQuery(interval)
 }
